backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pymysql>=1.1.0
pydantic>=2.5.0
httpx>=0.25.2
python-dotenv>=1.0.0
cloudinary>=1.36.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
email-validator>=2.1.0
Pillow>=10.1.0
aiofiles>=23.2.1
jinja2>=3.1.2
google-generativeai>=0.3.0
